+++ packages/codenic_exception_converter/lib/codenic_exception_converter.dart
@@ -2,7 +2,8 @@ export 'package:codenic_logger/codenic_logger.dart';
 export 'package:equatable/equatable.dart';
 export 'package:fpdart/fpdart.dart' show Either, Left, Right;
 
+export 'src/base_converters/error_converter.dart';
+export 'src/base_converters/exception_converter.dart';
 export 'src/exception_converter_suite.dart';
-export 'src/exception_converters/exception_converter.dart';
 export 'src/failures/failure.dart';
 export 'src/util/either_extension.dart';